+++ minhas-financas-back-end/MinhasFinancas.Application/Interfaces/IMfScoreLaboratorioAppService.cs
@@ -1,0 +1,8 @@
+namespace MinhasFinancas.Application.Interfaces
+{
+    public interface IMfScoreLaboratorioAppService
+    {
+        Task<RetornoGenerico> BuscarUsuariosAsync();
+        Task<RetornoGenerico> BuscarScoreUsuarioAsync(string usuarioId);
+    }
+}