+++ minhas-financas-back-end/MinhasFinancas.Infra/Data/Interfaces/IUsuarioRepository.cs
@@ -6,6 +6,8 @@ namespace MinhasFinancas.Infra.Data.Interfaces
     {
         Task<bool> ExisteUsuarioAsync(string usuarioId);
         Task<List<string>> BuscarIdsUsuariosAtivosAsync();
+        Task<List<Usuario>> BuscarUsuariosParaLaboratorioAsync();
+        Task<Usuario?> BuscarResumoUsuarioAsync(string usuarioId);
         Task DeletarUsuarioESeusDados(Usuario elemento);
     }
 }